fix(quality): resolve duplicated authjs strategy in Users + AnonymousUsers
Extracts the NextAuth-backed Payload strategy and the /logout endpoint
into a shared module so both collections call the same implementation.
Brings new-code duplicated lines density back under the 3% quality gate
threshold without changing runtime behavior.
